refactor: 피드백 적용
- 인수 테스트 Profile 통일
- 인수 테스트 공통 부분 부모 클래스로 분리
- dto 객체 request/respose 네이밍 컨벤션 통일
